requirements.txt
langchain-openai>=0.0.2
python-dotenv>=1.0.0
browser-use>=0.1.0
openai>=1.3.0
asyncio>=3.4.3
supabase>=2.3.4
numpy>=1.26.2
